# server/swagger.json
{
	"components": {
...
		"schemas": {
			"QuizOption": {
				"properties": {
					"text": {
						"type": "string"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"text",
					"id"
				],
				"type": "object"
			},
			"QuizResponse": {
				"properties": {
					"quizId": {
						"type": "string"
					},
					"quizCompleted": {
						"type": "boolean"
					},
					"previouslyQuestion_CorrectOptionId": {
						"type": "string"
					},
					"options": {
						"items": {
							"$ref": "#/components/schemas/QuizOption"
						},
						"type": "array"
					},
					"questionId": {
						"type": "string"
					},
					"questionText": {
						"type": "string"
					}
				},
				"required": [
					"options",
					"questionId",
					"questionText"
				],
				"type": "object"
			},
			"QuizRequestBody": {
				"properties": {
					"optionId": {
						"type": "string"
					},
					"questionId": {
						"type": "string"
					}
				},
				"required": [
					"optionId",
					"questionId"
				],
				"type": "object"
			},
			"EpilogueOption": {
				"properties": {
					"text": {
						"type": "string"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"text",
					"id"
				],
				"type": "object"
			},
			"EpilogueQuestion": {
				"properties": {
					"options": {
						"items": {
							"$ref": "#/components/schemas/EpilogueOption"
						},
						"type": "array"
					},
					"text": {
						"type": "string"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"options",
					"text",
					"id"
				],
				"type": "object"
			},
			"Epilogue": {
				"properties": {
					"questions": {
						"items": {
							"$ref": "#/components/schemas/EpilogueQuestion"
						},
						"type": "array"
					},
					"textStoryTale": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"questions",
					"textStoryTale",
					"name",
					"id"
				],
				"type": "object"
			},
			"EpilogueQuestionProgress": {
				"properties": {
					"question": {
						"$ref": "#/components/schemas/EpilogueQuestion"
					},
					"userStoryId": {
						"type": "string"
					},
					"questionId": {
						"type": "string"
					},
					"order": {
						"type": "number",
						"format": "double"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"userStoryId",
					"questionId",
					"order",
					"id"
				],
				"type": "object"
			},
			"EpilogueProgress": {
				"properties": {
					"timeCompleted": {
						"type": "number",
						"format": "double"
					},
					"timeStarted": {
						"type": "number",
						"format": "double"
					},
					"timeUnlocked": {
						"type": "number",
						"format": "double"
					},
					"timeSummaryCompleted": {
						"type": "number",
						"format": "double"
					},
					"questionProgressItems": {
						"properties": {},
						"additionalProperties": {
							"$ref": "#/components/schemas/EpilogueQuestionProgress"
						},
						"type": "object"
					},
					"lessonStoryId": {
						"type": "string"
					},
					"userStoryId": {
						"type": "string"
					},
					"epilogue": {
						"$ref": "#/components/schemas/Epilogue"
					},
					"epilogueId": {
						"type": "string"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"questionProgressItems",
					"lessonStoryId",
					"userStoryId",
					"epilogueId",
					"id"
				],
				"type": "object"
			},
			"Word": {
				"properties": {
					"longTranslation": {
						"type": "string"
					},
					"shortTranslation": {
						"type": "string"
					},
					"text": {
						"type": "string"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"longTranslation",
					"shortTranslation",
					"text",
					"id"
				],
				"type": "object"
			},
			"BuildingBlock": {
				"properties": {
					"words": {
						"items": {
							"$ref": "#/components/schemas/Word"
						},
						"type": "array"
					},
					"isStarter": {
						"type": "boolean"
					},
					"idsItemsDependentOnThis": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"imageUrl": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"words",
					"isStarter",
					"imageUrl",
					"name",
					"id"
				],
				"type": "object"
			},
			"WordProgress": {
				"properties": {
					"word": {
						"$ref": "#/components/schemas/Word"
					},
					"wordId": {
						"type": "string"
					},
					"userStoryId": {
						"type": "string"
					},
					"order": {
						"type": "number",
						"format": "double"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"wordId",
					"userStoryId",
					"order",
					"id"
				],
				"type": "object"
			},
			"BuildingBlockProgress": {
				"properties": {
					"timeSummaryCompleted": {
						"type": "number",
						"format": "double"
					},
					"timeCompleted": {
						"type": "number",
						"format": "double"
					},
					"timeStarted": {
						"type": "number",
						"format": "double"
					},
					"timeUnlocked": {
						"type": "number",
						"format": "double"
					},
					"isStarter": {
						"type": "boolean"
					},
					"wordProgressItems": {
						"properties": {},
						"additionalProperties": {
							"$ref": "#/components/schemas/WordProgress"
						},
						"type": "object"
					},
					"lessonStoryId": {
						"type": "string"
					},
					"userStoryId": {
						"type": "string"
					},
					"isDependentOnNames": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"block": {
						"$ref": "#/components/schemas/BuildingBlock"
					},
					"blockId": {
						"type": "string"
					},
					"order": {
						"type": "number",
						"format": "double"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"isStarter",
					"wordProgressItems",
					"lessonStoryId",
					"userStoryId",
					"blockId",
					"order",
					"id"
				],
				"type": "object"
			},
			"QuizBlockCompletedStatsResponse": {
				"properties": {
					"blockCompletedStoryRefId": {
						"type": "string"
					},
					"blockCompleted": {
						"$ref": "#/components/schemas/BuildingBlockProgress"
					},
					"blockProgressUnlockedItems": {
						"items": {
							"$ref": "#/components/schemas/BuildingBlockProgress"
						},
						"type": "array"
					},
					"epilogueProgressUnlocked": {
						"$ref": "#/components/schemas/EpilogueProgress"
					}
				},
				"required": [
					"blockCompletedStoryRefId"
				],
				"type": "object"
			},
...
			"Pick_EpilogueProgress.Exclude_keyofEpilogueProgress.questionProgressItems__": {
				"properties": {
					"id": {
						"type": "string"
					},
					"userStoryId": {
						"type": "string"
					},
					"lessonStoryId": {
						"type": "string"
					},
					"timeUnlocked": {
						"type": "number",
						"format": "double"
					},
					"timeStarted": {
						"type": "number",
						"format": "double"
					},
					"timeCompleted": {
						"type": "number",
						"format": "double"
					},
					"timeSummaryCompleted": {
						"type": "number",
						"format": "double"
					},
					"epilogueId": {
						"type": "string"
					},
					"epilogue": {
						"$ref": "#/components/schemas/Epilogue"
					}
				},
				"required": [
					"id",
					"userStoryId",
					"lessonStoryId",
					"epilogueId"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_EpilogueProgress.questionProgressItems_": {
				"$ref": "#/components/schemas/Pick_EpilogueProgress.Exclude_keyofEpilogueProgress.questionProgressItems__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"EpilogueProgressOutput": {
				"allOf": [
					{
						"$ref": "#/components/schemas/Omit_EpilogueProgress.questionProgressItems_"
					},
					{
						"properties": {
							"questionProgressItems": {
								"items": {
									"$ref": "#/components/schemas/EpilogueQuestionProgress"
								},
								"type": "array"
							}
						},
						"required": [
							"questionProgressItems"
						],
						"type": "object"
					}
				]
			},
...
		},
		"securitySchemes": {
			"BearerAuth": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header",
				"description": "Bearer token"
			}
		}
	},
	"info": {
		"title": "lang-learning-app-server",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/api/blocks/{blockProgressId}/quiz/request-question": {
			"post": {
				"operationId": "RequestQuizQuestion",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/QuizResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Quiz - Builiding blocks"
				],
				"security": [
					{
						"BearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "blockProgressId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/blocks/{blockProgressId}/quiz/answer-question": {
			"post": {
				"operationId": "AnswerQuizQuestion",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/QuizResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Quiz - Builiding blocks"
				],
				"security": [
					{
						"BearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "blockProgressId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/QuizRequestBody"
							}
						}
					}
				}
			}
		},
		"/api/blocks/{blockProgressId}/quiz/{quizId}/completed": {
			"get": {
				"operationId": "GetProgressAchievedOfCompletedQuiz",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/QuizBlockCompletedStatsResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Quiz - Builiding blocks"
				],
				"security": [
					{
						"BearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "blockProgressId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "quizId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/api/epilogues/{epilogueProgressId}": {
			"get": {
				"operationId": "GetEpilogueProgress",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/EpilogueProgressOutput"
								}
							}
						}
					}
				},
				"tags": [
					"Epilogue blocks"
				],
				"security": [
					{
						"BearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "epilogueProgressId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/api/epilogues/{epilogueProgressId}/quiz/request-question": {
			"post": {
				"operationId": "RequestQuizQuestion",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/QuizResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Quiz - Epilogue blocks"
				],
				"security": [
					{
						"BearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "epilogueProgressId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/epilogues/{epilogueProgressId}/quiz": {
			"post": {
				"operationId": "AnswerQuizQuestion",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/QuizResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Quiz - Epilogue blocks"
				],
				"security": [
					{
						"BearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "epilogueProgressId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/QuizRequestBody"
							}
						}
					}
				}
			}
		},
		"/api/epilogues/{epilogueProgressId}/quiz/{quizId}/completed": {
			"get": {
				"operationId": "GetProgressAchievedOfCompletedQuiz",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/QuizBlockCompletedStatsResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Quiz - Epilogue blocks"
				],
				"security": [
					{
						"BearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "epilogueProgressId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "quizId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}